refactor: Move message.dart to models/chat
