+++ lib/features/mode_demo/mode_demo_chat_repository.dart
@@ -1,4 +1,4 @@
-import 'package:pass_emploi_app/models/message.dart';
+import 'package:pass_emploi_app/models/chat/message.dart';
 import 'package:uuid/uuid.dart';
 
 List<Message> modeDemoChat() => [
