+++ lib/models/chat/message.dart
