+++ lib/models/chat/offre_partagee.dart
@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:pass_emploi_app/models/message.dart';
+import 'package:pass_emploi_app/models/chat/message.dart';
 
 class OffrePartagee extends Equatable {
   final String id;
