+++ lib/presentation/chat/chat_item.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
-import 'package:pass_emploi_app/models/message.dart';
+import 'package:pass_emploi_app/models/chat/message.dart';
 
 sealed class ChatItem extends Equatable {
   final String messageId;
